refactor: implement minFreeSpace function for Unix and Windows environments
